+++ NEWS.md
@@ -1,3 +1,7 @@
+# rostemplate (development version)
+
+-   Use `testthat` package.
+
 # rostemplate 0.0.3
 
 -   Add `ros_badge_ropenspain()` : [![rOS-badge](https://ropenspain.github.io/rostemplate/reference/figures/ropenspain-badge.svg)](https://ropenspain.es/)